+++ PassMan/Policy/PolicyException.php
@@ -1,9 +1,11 @@
 <?php
+
+declare(strict_types=1);
+
 namespace Vanqard\PassMan\Policy;
 
 use Exception;
 
 class PolicyException extends Exception
 {
-
 }